Move the TokenCredential trait to azure_core
