+++ sdk/identity/examples/cli_credentials.rs
@@ -1,3 +1,4 @@
+use azure_core::TokenCredential;
 use azure_identity::token_credentials::*;
 use std::error::Error;
 use url::Url;
