+++ sdk/identity/src/token_credentials/cli_credentials.rs
@@ -1,6 +1,5 @@
-use crate::token_credentials::{TokenCredential, TokenResponse};
-
 use azure_core::errors::AzureError;
+use azure_core::{TokenCredential, TokenResponse};
 use chrono::{DateTime, Utc};
 use oauth2::AccessToken;
 use serde::Deserialize;
